[LED-66] remove FrequencyConverter and cleanup code
