+++ src/main/java/ledger/database/storage/SQL/ISQLDatabaseFrequency.java
@@ -1,7 +1,6 @@
 package ledger.database.storage.SQL;
 
 import ledger.database.entity.Frequency;
-import ledger.database.entity.FrequencyConverter;
 import ledger.database.storage.SQL.SQLite.ISQLiteDatabase;
 import ledger.database.storage.table.FrequencyTable;
 import ledger.exception.StorageException;
@@ -22,7 +21,7 @@ public interface ISQLDatabaseFrequency extends ISQLiteDatabase {
         try {
             PreparedStatement stmt = getDatabase().prepareStatement("INSERT INTO " + FrequencyTable.TABLE_NAME +
                     " (" + FrequencyTable.FREQUENCY_NAME + ") VALUES (?)");
-            stmt.setString(1, FrequencyConverter.convertFrequencyToString(frequency));
+            stmt.setString(1, Frequency.convertFrequencyToString(frequency));
             stmt.executeUpdate();
         } catch (java.sql.SQLException e) {
             throw new StorageException("Error while inserting frequencies from database.", e);
@@ -38,7 +37,7 @@ public interface ISQLDatabaseFrequency extends ISQLiteDatabase {
 
             while (rs.next()) {
                 String frequencyName = rs.getString(FrequencyTable.FREQUENCY_NAME);
-                Frequency currentFrequency = FrequencyConverter.convertStringToFrequency(frequencyName);
+                Frequency currentFrequency = Frequency.convertStringToFrequency(frequencyName);
                 frequencies.add(currentFrequency);
             }
         } catch (java.sql.SQLException e) {
@@ -53,7 +52,7 @@ public interface ISQLDatabaseFrequency extends ISQLiteDatabase {
             PreparedStatement stmt = getDatabase().prepareStatement("SELECT " + FrequencyTable.FREQUENCY_ID +
                     " FROM " + FrequencyTable.TABLE_NAME +
                     " WHERE " + FrequencyTable.FREQUENCY_NAME + "=?");
-            stmt.setString(1, FrequencyConverter.convertFrequencyToString(frequency));
+            stmt.setString(1, Frequency.convertFrequencyToString(frequency));
 
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
@@ -76,7 +75,7 @@ public interface ISQLDatabaseFrequency extends ISQLiteDatabase {
 
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
-                toReturn = FrequencyConverter.convertStringToFrequency(rs.getString(FrequencyTable.FREQUENCY_NAME));
+                toReturn = Frequency.convertStringToFrequency(rs.getString(FrequencyTable.FREQUENCY_NAME));
             }
 
             return toReturn;
